+++ maintenance-core/src/main/java/eu/kennytv/maintenance/core/MaintenancePlugin.java
@@ -26,6 +26,7 @@ import eu.kennytv.maintenance.api.IMaintenance;
 import eu.kennytv.maintenance.api.event.MaintenanceChangedEvent;
 import eu.kennytv.maintenance.api.event.manager.IEventManager;
 import eu.kennytv.maintenance.core.command.MaintenanceCommand;
+import eu.kennytv.maintenance.core.config.ConfigSection;
 import eu.kennytv.maintenance.core.dump.MaintenanceDump;
 import eu.kennytv.maintenance.core.dump.PluginDump;
 import eu.kennytv.maintenance.core.event.EventManager;
@@ -44,6 +45,7 @@ import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 import java.util.List;
 import java.util.UUID;
+import java.util.concurrent.TimeUnit;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -57,13 +59,23 @@ public abstract class MaintenancePlugin implements IMaintenance {
     private final String prefix;
     private final ServerType serverType;
     private Version newestVersion;
-    private Task task;
 
     protected MaintenancePlugin(final String version, final ServerType serverType) {
         this.version = new Version(version);
         this.serverType = serverType;
         this.prefix = "§8[§eMaintenance" + serverType + "§8] ";
         this.eventManager = new EventManager();
+    }
+
+    public void disable() {
+        // Save the current endtimer to continue it after the next restart
+        if (runnable != null && !runnable.shouldEnable()) {
+            final ConfigSection section = settings.getConfig().getSection("continue-endtimer-after-restart");
+            if (section != null) {
+                section.set("end", System.currentTimeMillis() + TimeUnit.SECONDS.toMillis(runnable.getSecondsLeft()));
+                settings.saveConfig();
+            }
+        }
     }
 
     @Override
@@ -99,14 +111,41 @@ public abstract class MaintenancePlugin implements IMaintenance {
                 .replace("%SECONDS%", String.format("%02d", seconds));
     }
 
-    public void startMaintenanceRunnable(final int minutes, final boolean enable) {
-        runnable = new MaintenanceRunnable(this, settings, minutes, enable);
-        task = startMaintenanceRunnable(runnable);
+    public void startMaintenanceRunnableForMinutes(final int minutes, final boolean enable) {
+        startMaintenanceRunnableForSeconds(minutes * 60, enable);
+    }
+
+    public void startMaintenanceRunnableForSeconds(final int seconds, final boolean enable) {
+        runnable = new MaintenanceRunnable(this, settings, seconds, enable);
     }
 
     public boolean updateAvailable() {
         checkNewestVersion();
         return version.compareTo(newestVersion) == -1;
+    }
+
+    protected void continueLastEndtimer() {
+        final ConfigSection section = settings.getConfig().getSection("continue-endtimer-after-restart");
+        if (section == null) return;
+        if (!section.getBoolean("enabled", false)) return;
+
+        final long end = section.getLong("end");
+        if (end == 0) return;
+
+        final long current = System.currentTimeMillis();
+        getLogger().info("Found interrupted endtimer from last uptime...");
+        if (end < current) {
+            getLogger().info("The endtimer has already expired, thus it has been cancelled.");
+        } else if (!isMaintenance()) {
+            getLogger().info("Maintenance has already been disabled, thus the timer has been cancelled.");
+        } else {
+            final int seconds = (int) TimeUnit.MILLISECONDS.toSeconds(end - current);
+            startMaintenanceRunnableForSeconds(seconds, false);
+            getLogger().info("The timer has been continued - maintenance will be disabled in: " + formatedTimer());
+        }
+
+        section.remove("end");
+        settings.saveConfig();
     }
 
     protected void sendEnableMessage() {
@@ -228,9 +267,8 @@ public abstract class MaintenancePlugin implements IMaintenance {
     }
 
     public void cancelTask() {
-        task.cancel();
+        runnable.getTask().cancel();
         runnable = null;
-        task = null;
     }
 
     public UUID checkUUID(final SenderInfo sender, final String s) {
@@ -255,7 +293,7 @@ public abstract class MaintenancePlugin implements IMaintenance {
 
     @Override
     public boolean isTaskRunning() {
-        return task != null && runnable != null;
+        return runnable != null;
     }
 
     @Override
@@ -307,6 +345,8 @@ public abstract class MaintenancePlugin implements IMaintenance {
 
     public abstract void sendUpdateNotification(SenderInfo sender);
 
+    public abstract Task startMaintenanceRunnable(Runnable runnable);
+
     public abstract SenderInfo getOfflinePlayer(String name);
 
     public abstract SenderInfo getOfflinePlayer(UUID uuid);
@@ -324,8 +364,6 @@ public abstract class MaintenancePlugin implements IMaintenance {
     protected abstract void loadIcon(File file) throws Exception;
 
     protected abstract void kickPlayers();
-
-    protected abstract Task startMaintenanceRunnable(Runnable runnable);
 
     protected abstract File getPluginFile();
 }