+++ maintenance-spigot/src/main/java/eu/kennytv/maintenance/spigot/MaintenanceSpigotPlugin.java
@@ -90,6 +90,7 @@ public final class MaintenanceSpigotPlugin extends MaintenancePlugin {
             }
         }
 
+        continueLastEndtimer();
         new MetricsLite(plugin);
 
         // ServerListPlus integration